feat(api): add getRecvFriendRequest API and update related types and resolve (#135)
- Introduced getRecvFriendRequest API for retrieving received friend requests, including request handling and response types.
- Updated index and Zalo class to integrate the new getRecvFriendRequest method.
- Simplified GetFriendRequestResponse type and added new GetRecvFriendRequestResponse type for better clarity and structure.
