+++ src/apis/getFriendRequest.ts
@@ -1,55 +1,29 @@
 import { ZaloApiError } from "../Errors/ZaloApiError.js";
 import { apiFactory } from "../utils.js";
 
-export type CollapseMsgListConfig = {
-    collapseId: number;
-    collapseXItem: number;
-    collapseYItem: number;
-};
-
-export type RecommInfo = {
-    source: number;
-    message: string;
-};
-
-export type BizPkg = {
-    pkgId: number;
-};
-
-export type DataInfo = {
-    userId: string;
-    zaloName: string;
-    displayName: string;
-    avatar: string;
-    phoneNumber: string;
-    status: string;
-    gender: number;
-    dob: number;
-    type: number;
-    recommType: number;
-    recommSrc: number;
-    recommTime: number;
-    recommInfo: RecommInfo;
-    bizPkg: BizPkg;
-    isSeenFriendReq: boolean;
-};
-
-export type RecommItem = {
-    recommItemType: number;
-    dataInfo: DataInfo;
-};
-
 export type GetFriendRequestResponse = {
-    expiredDuration: number;
-    collapseMsgListConfig: CollapseMsgListConfig;
-    recommItems: RecommItem[];
+    [key: string]: {
+        userId: string;
+        zaloName: string;
+        displayName: string;
+        avatar: string;
+        globalId: string;
+        bizPkg: {
+            pkgId: number;
+        };
+        fReqInfo: {
+            message: string;
+            src: number;
+            time: number;
+        };
+    };
 };
 
 export const getFriendRequestFactory = apiFactory<GetFriendRequestResponse>()((api, ctx, utils) => {
-    const serviceURL = utils.makeURL(`${api.zpwServiceMap.friend[0]}/api/friend/recommendsv2/list`);
+    const serviceURL = utils.makeURL(`${api.zpwServiceMap.friend[0]}/api/friend/requested/list`);
 
     /**
-     * Get friend requests
+     * Get friend requested
      *
      * @throws ZaloApiError
      */
